Правки модулей
