+++ js/main.js
@@ -1,2 +1,5 @@
 import { SIMILAR_OBJECT } from './data';
+import { checkStringLength } from './utils';
 SIMILAR_OBJECT();
+checkStringLength('Hello world', 10);
+
